+++ src/y2019/day5/MachIO.java
@@ -3,22 +3,22 @@ package y2019.day5;
 import y2019.intcode.Machine;
 
 public class MachIO implements Machine.IO {
-    private long nextInput;
-    private long lastOutput;
+    private int nextInput;
+    private int lastOutput;
 
-    public MachIO(long value) {
+    public MachIO(int value) {
         this.nextInput = value;
     }
 
-    public long input() {
+    public int input() {
         return nextInput;
     }
 
-    public void output(long value) {
+    public void output(int value) {
         lastOutput = value;
     }
 
-    public long getLastOutput() {
+    public int getLastOutput() {
         return lastOutput;
     }
 }
